+++ framework/Stdio.php
@@ -6,8 +6,10 @@ namespace Ssc\Btlr\Framework;
 
 use Ssc\Btlr\Framework\Stdio\Write;
 use Ssc\Btlr\Framework\Stdio\Write\WithStyle;
+use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\Question;
 
 class Stdio
 {
@@ -17,7 +19,17 @@ class Stdio
         private InputInterface $input,
         private OutputInterface $output,
         private Write $write = new Write(),
+        private QuestionHelper $questionHelper = new QuestionHelper(),
     ) {
+    }
+
+    public function ask(string $question): string
+    {
+        return $this->questionHelper->ask(
+            $this->input,
+            $this->output,
+            new Question("<fg=green>{$question}</>"),
+        );
     }
 
     public function write(