+++ framework/Time/Clock.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ssc\Btlr\Framework\Time;
+
+class Clock
+{
+    public function inFormat(string $format): string
+    {
+        return (new \DateTime())->format($format);
+    }
+}